+++ CHANGES.md
@@ -25,6 +25,8 @@ with `allowNegativeBalance` as `true`.
     and `SetValidator()` of `IWorld` to extension methods.  [[#3715]]
  -  (Libplanet.Action) Removed `allowNegativeBalance` parameter from
     `IWorld.TransferAsset()` extension method.  [[#3725], [#3728]]
+ -  (Libplanet.Store) Removed `journal`, `indexCacheSize`, and `flush`
+    parameters from `DefaultStore`'s constructor.  [[#3729]]
 
 ### Backward-incompatible network protocol changes
 
@@ -38,6 +40,9 @@ with `allowNegativeBalance` as `true`.
 
 ### Dependencies
 
+ -  Upgraded *LiteDB* from [4.1.4][LiteDB 4.1.4] to
+    [5.0.15][LiteDB 5.0.15].  [[#3729]]
+
 ### CLI tools
 
 [#3713]: https://github.com/planetarium/libplanet/pull/3713
@@ -45,6 +50,9 @@ with `allowNegativeBalance` as `true`.
 [#3715]: https://github.com/planetarium/libplanet/pull/3715
 [#3725]: https://github.com/planetarium/libplanet/issues/3725
 [#3728]: https://github.com/planetarium/libplanet/pull/3728
+[#3729]: https://github.com/planetarium/libplanet/pull/3729
+[LiteDB 4.1.4]: https://www.nuget.org/packages/LiteDB/4.1.4
+[LiteDB 5.0.15]: https://www.nuget.org/packages/LiteDB/5.0.15
 
 
 Version 4.2.0